Simplifying mock data classes
