+++ app/lib/features/chat_ui_showcase/mocks/convo/timeline/mock_timeline_stream.dart
@@ -1,45 +1,6 @@
-import 'package:acter/features/chat_ui_showcase/models/mocks/convo/mock_convo.dart';
+import 'package:acter/features/chat_ui_showcase/mocks/convo/timeline/mock_timeline_item_diff.dart';
 import 'package:acter_flutter_sdk/acter_flutter_sdk_ffi.dart';
 import 'package:mocktail/mocktail.dart';
-
-class MockFfiListTimelineItem extends Mock implements FfiListTimelineItem {
-  final List<MockTimelineItem> timelineItems;
-
-  MockFfiListTimelineItem({required this.timelineItems});
-
-  @override
-  List<MockTimelineItem> toList({bool growable = false}) =>
-      timelineItems.toList(growable: growable);
-}
-
-class MockTimelineItemDiff extends Mock implements TimelineItemDiff {
-  final String mockAction;
-  final MockFfiListTimelineItem? mockTimelineItemList;
-  final int? mockIndex;
-  final MockTimelineItem? mockTimelineItem;
-
-  MockTimelineItemDiff({
-    this.mockAction = 'Append',
-    this.mockTimelineItemList,
-    this.mockIndex,
-    this.mockTimelineItem,
-  });
-
-  @override
-  String action() => mockAction;
-
-  @override
-  FfiListTimelineItem? values() => mockTimelineItemList;
-
-  @override
-  int? index() => mockIndex;
-
-  @override
-  TimelineItem? value() => mockTimelineItem;
-
-  @override
-  void drop() {}
-}
 
 class MockTimelineStream extends Mock implements TimelineStream {
   final List<MockTimelineItemDiff> mockTimelineItemDiffs;
