+++ app/lib/features/chat_ui_showcase/mocks/general/mock_msg_content.dart
@@ -1,0 +1,10 @@
+import 'package:acter_flutter_sdk/acter_flutter_sdk_ffi.dart';
+import 'package:mocktail/mocktail.dart';
+
+class MockMsgContent extends Mock implements MsgContent {
+  final String? mockBody;
+  MockMsgContent({this.mockBody});
+
+  @override
+  String body() => mockBody ?? 'body';
+}
