+++ app/lib/features/chat_ui_showcase/mocks/general/mock_option_string.dart
@@ -1,0 +1,11 @@
+import 'package:acter_flutter_sdk/acter_flutter_sdk_ffi.dart';
+import 'package:mocktail/mocktail.dart';
+
+class MockOptionString extends Mock implements OptionString {
+  final String? mockText;
+
+  MockOptionString({this.mockText});
+
+  @override
+  String? text() => mockText;
+}
