+++ app/lib/features/chat_ui_showcase/mocks/showcase_data/chat_event_showcase_list.dart
@@ -1,4 +1,5 @@
-import 'package:acter/features/chat_ui_showcase/models/mocks/convo/mock_convo.dart';
+import 'package:acter/features/chat_ui_showcase/mocks/convo/timeline/mock_timeline_event_item.dart';
+import 'package:acter/features/chat_ui_showcase/mocks/convo/timeline/mock_timeline_item.dart';
 
 List<MockTimelineItem> mockChatEventList = [
   MockTimelineItem(
