+++ app/lib/features/chat_ui_showcase/pages/chat_list_showcase_page.dart
@@ -2,7 +2,7 @@ import 'package:acter/common/utils/routes.dart';
 import 'package:acter/common/widgets/plus_icon_widget.dart';
 import 'package:acter/common/toolkit/widgets/animated_chats_list_widget.dart';
 import 'package:acter/features/chat_ng/rooms_list/widgets/chat_item_widget.dart';
-import 'package:acter/features/chat_ui_showcase/models/mocks/showcase_data/convo_showcase_list.dart';
+import 'package:acter/features/chat_ui_showcase/mocks/showcase_data/convo_showcase_list.dart';
 import 'package:acter/l10n/generated/l10n.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
