+++ app/test/features/chat_ng/widgets/chat_item/last_message_widgets/golden_profile_changes_message_event_test.dart
@@ -1,5 +1,5 @@
 import 'package:acter/features/chat_ng/rooms_list/widgets/chat_item_widget.dart';
-import 'package:acter/features/chat_ui_showcase/models/mocks/showcase_data/convo_showcase_data.dart';
+import 'package:acter/features/chat_ui_showcase/mocks/showcase_data/convo_showcase_data.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import '../../../../../helpers/font_loader.dart';
