+++ app/test/features/chat_ng/widgets/chat_item/last_message_widgets/room_membership_event_widget_test.dart
@@ -1,8 +1,8 @@
 import 'package:acter/common/providers/common_providers.dart';
 import 'package:acter/features/chat_ng/providers/chat_list_providers.dart';
 import 'package:acter/features/chat_ng/widgets/chat_item/last_message_widgets/room_membership_event_widget.dart';
-import 'package:acter/features/chat_ui_showcase/models/mocks/showcase_data/convo_showcase_data.dart';
-import 'package:acter/features/chat_ui_showcase/models/mocks/convo/mock_convo.dart';
+import 'package:acter/features/chat_ui_showcase/mocks/showcase_data/convo_showcase_data.dart';
+import 'package:acter/features/chat_ui_showcase/mocks/convo/mock_convo.dart';
 import 'package:acter/l10n/generated/l10n.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
